% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_index}
\alias{calculate_index}
\title{Main function to calculate the spectral index based on an XML definition.}
\usage{
calculate_index(xml_file, wavelengths, reflectance_list)
}
\description{
Main function to calculate the spectral index based on an XML definition.
}
