% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scancorder_multi_calibration.R
\docType{class}
\name{CalibrationReflectanceMultipoint}
\alias{CalibrationReflectanceMultipoint}
\title{CalibrationReflectanceMultipoint R6 Class}
\description{
Performs multi-point reflectance calibration based on a 3D array of
calibration factors, typically loaded from JSON data from CICADA sensor
reading.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalibrationReflectanceMultipoint-new}{\code{CalibrationReflectanceMultipoint$new()}}
\item \href{#method-CalibrationReflectanceMultipoint-slice_keep_first}{\code{CalibrationReflectanceMultipoint$slice_keep_first()}}
\item \href{#method-CalibrationReflectanceMultipoint-multi_point_calibration}{\code{CalibrationReflectanceMultipoint$multi_point_calibration()}}
\item \href{#method-CalibrationReflectanceMultipoint-convert_json_to_3d_array}{\code{CalibrationReflectanceMultipoint$convert_json_to_3d_array()}}
\item \href{#method-CalibrationReflectanceMultipoint-score}{\code{CalibrationReflectanceMultipoint$score()}}
\item \href{#method-CalibrationReflectanceMultipoint-clone}{\code{CalibrationReflectanceMultipoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibrationReflectanceMultipoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalibrationReflectanceMultipoint-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibrationReflectanceMultipoint$new(calibration_factors = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibrationReflectanceMultipoint-slice_keep_first"></a>}}
\if{latex}{\out{\hypertarget{method-CalibrationReflectanceMultipoint-slice_keep_first}{}}}
\subsection{Method \code{slice_keep_first()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibrationReflectanceMultipoint$slice_keep_first(x, k)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibrationReflectanceMultipoint-multi_point_calibration"></a>}}
\if{latex}{\out{\hypertarget{method-CalibrationReflectanceMultipoint-multi_point_calibration}{}}}
\subsection{Method \code{multi_point_calibration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibrationReflectanceMultipoint$multi_point_calibration(sensor_values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibrationReflectanceMultipoint-convert_json_to_3d_array"></a>}}
\if{latex}{\out{\hypertarget{method-CalibrationReflectanceMultipoint-convert_json_to_3d_array}{}}}
\subsection{Method \code{convert_json_to_3d_array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibrationReflectanceMultipoint$convert_json_to_3d_array(
  json_data,
  type = as.numeric
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibrationReflectanceMultipoint-score"></a>}}
\if{latex}{\out{\hypertarget{method-CalibrationReflectanceMultipoint-score}{}}}
\subsection{Method \code{score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibrationReflectanceMultipoint$score(reflectance, json_input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibrationReflectanceMultipoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalibrationReflectanceMultipoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibrationReflectanceMultipoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
