% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_mathml.R
\name{evaluate_mathml}
\alias{evaluate_mathml}
\title{Recursive function to evaluate MathML expressions.
This function expects an XML node and a named list 'values' with numeric values.}
\usage{
evaluate_mathml(node, values)
}
\arguments{
\item{node}{XML node representing a MathML element.}

\item{values}{Named list of variable identifiers and their numeric values.}
}
\value{
Numeric value resulting from the evaluation of the MathML expression.
}
\description{
Recursive function to evaluate MathML expressions.
This function expects an XML node and a named list 'values' with numeric values.
}
