% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sensor_metadata.R
\name{find_sensor_metadata}
\alias{find_sensor_metadata}
\title{Find sensor metadata by a 4-digit serial embedded in sensor_serial}
\usage{
find_sensor_metadata(
  serial_name,
  directory = system.file("extdata", "sensors", package = "scancorder.indices")
)
}
\arguments{
\item{serial_name}{A 4-digit serial (character or numeric)}

\item{directory}{Path to folder containing .json files.
Defaults to the "sensors" sub directory of the installed scancorder.indices package.}
}
\value{
Parsed JSON (as an R list) for the first file whose sensor_serial contains that 4-digit code, or NULL
}
\description{
Find sensor metadata by a 4-digit serial embedded in sensor_serial
}
