% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_names.R
\name{get_index_names}
\alias{get_index_names}
\title{Extract the \\<Name\\> element from spectral-index XML files in the package}
\usage{
get_index_names()
}
\value{
A character vector of index names.
}
\description{
This function automatically discovers XML files in the package's indices directory
and extracts the index names from each file.
}
