% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_xml.R
\name{get_index_xml}
\alias{get_index_xml}
\title{Find the file path to a spectral-index XML by name from the scancorder.indices package}
\usage{
get_index_xml(index_name)
}
\arguments{
\item{index_name}{Character. Name of the index to find (e.g. "NDVI").}
}
\value{
Character. Full file path to the matching XML.
}
\description{
This function locates XML files shipped in the \verb{inst/indices/} folder of the
\code{scancorder.indices} package and returns the full file path of the XML whose
\\<Name\\> element matches \code{index_name}.
}
