% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_names.R
\name{get_index_names}
\alias{get_index_names}
\title{Extract the \if{html}{\out{<Name>}} element from one or more spectral-index XML files}
\usage{
get_index_names()
}
\arguments{
\item{xml_files}{Character vector with paths to the XML files \strong{or}
a single directory that contains only index XMLs.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A character vector of index names.
}\if{html}{\out{</div>}}
}
\description{
Extract the \if{html}{\out{<Name>}} element from one or more spectral-index XML files
}
\examples{
files <- list.files("indices", pattern = "\\\\.xml$", full.names = TRUE)
get_index_names(files)
#> [1] "NDVI" "NDWI"
}
